Adjusting additional serializers.
