+++ api_v2/serializers/__init__.py
@@ -1,6 +1,6 @@
 """The initialization for serializers for open5e's api v2."""
 
-from .serializers import *
+#from .serializers import *
 
 from .item import ArmorSerializer
 from .item import WeaponSerializer
@@ -9,3 +9,16 @@ from .item import ItemSetSerializer
 
 from .background import BackgroundBenefitSerializer
 from .background import BackgroundSerializer
+
+from .document import RulesetSerializer
+from .document import LicenseSerializer
+from .document import PublisherSerializer
+from .document import DocumentSerializer
+
+from .feat import FeatBenefitSerializer
+from .feat import FeatSerializer
+
+from .race import TraitSerializer
+from .race import RaceSerializer
+
+from .creature import CreatureSerializer
